+++ Source/loadsave.cpp
@@ -212,7 +212,7 @@ public:
 void LoadItemData(LoadHelper &file, Item &item)
 {
 	item._iSeed = file.NextLE<int32_t>();
-	item._iCreateInfo = file.NextLE<uint16_t>();
+	item._iCreateInfo = file.NextLE<uint32_t>();
 	file.Skip(2); // Alignment
 	item._itype = static_cast<ItemType>(file.NextLE<uint32_t>());
 	item.position.x = file.NextLE<int32_t>();
@@ -936,7 +936,7 @@ void SaveItem(SaveHelper &file, const Item &item)
 	}
 
 	file.WriteLE<int32_t>(item._iSeed);
-	file.WriteLE<int16_t>(item._iCreateInfo);
+	file.WriteLE<int32_t>(item._iCreateInfo);
 	file.Skip(2); // Alignment
 	file.WriteLE<int32_t>(static_cast<int32_t>(iType));
 	file.WriteLE<int32_t>(item.position.x);
